== The check of 'unlikeky arguemtn type for collection methids using Object'    is set to Error with strict anylyses on compilation.    Hence some compiler errors must be adjusted, which are propblems before. * JZtxtcmdExecute, JZtxtcmdScript, FileRemote, XmlJzReader:   some implicitely 'toString()' now explicit because key is String * TreeNodeBase#detach(): explicitely cast to key type 'DerivedNode'. * IndexMultiTable.Set#contains(): here 'IndexMultiTable.this.containsValue(o);'   has returned anyway false, because it is not used. Now false immediately.   Problem is not really fixed but not relevant. * IndexMultiTableInteger#get(): correct cast to key type Integer now explicitely ==OutTextPreparer supports also 'gzTxt' as keyword

+++ java/org/vishia/util/IndexMultiTableInteger.java
@@ -727,7 +727,7 @@ public class IndexMultiTableInteger<Type> implements Map<Integer,Type>, Iterable
       else if(idx < sizeBlock)
       { child = ((IndexMultiTableInteger<Type>)(obj[idx]));
         //the child has space.
-        lastObj = child.get(key2); 
+        lastObj = child.get((Integer)key2); 
       }
     }
     else
